Update DaplugSession.java
According to the product specification, we should add the apdu Le parameter in the "Read binary" raw command. Le used instead of Lc, specifies the number of data that should be returned by the card.
But, in our approch, this is seen always as the Lc parameter. So, to prevent length errors, we add Lc zero-byte as fake data.

Without a SC, I found that the "Read binary" raw command returns a SW 6702 when we use a Le greather then the file size.. However, it works properly over a secure channel. To keep only one behaviour for both cases, I check the requested length value if it exceeds the file size.